pyproject.toml
[project]
name = "googlespannerpractice"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "google-cloud-spanner>=3.63.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
]
